(chore) move 3.4 test back into normal test suite
